feat: deploy on CI
