+++ .github/workflows/deploy.yml
@@ -1,0 +1,102 @@
+name: Deploy a stack
+on: [push]
+
+env:
+    ECR_REGISTRY: public.ecr.aws
+    ECR_REPOSITORY: succinct-labs/sp1-tee-enclave
+
+jobs:
+    docker-build-push:
+        runs-on:
+            [
+                "runs-on",
+                "runner=32cpu-linux-x64",
+                "run-id=${{ github.run_id }}",
+                "spot=false",
+            ]
+        steps:
+            - name: Retrieve the tag
+              run: |
+                  # echo "RELEASE_TAG=`git describe --tags --exact-match 2> /dev/null`" >> $GITHUB_ENV
+                  echo "ENCLAVE_VERSION=v1" >> $GITHUB_ENV
+                  echo "RELEASE_TAG=v1" >> $GITHUB_ENV
+
+            - name: Checkout repo
+              uses: actions/checkout@v4
+              with:
+                  fetch-depth: 0
+
+            - name: Configure AWS credentials
+              uses: "aws-actions/configure-aws-credentials@v1"
+              with:
+                  aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID }}
+                  aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+                  aws-region: ${{ secrets.AWS_REGION }}
+
+            - name: Set up Docker
+              uses: docker/setup-buildx-action@v3
+
+            - name: Login to Amazon ECR Public
+              id: login-ecr-public
+              uses: aws-actions/amazon-ecr-login@v2
+              with:
+                  registry-type: public
+
+            - name: Build Docker image
+              uses: docker/build-push-action@v6
+              with:
+                  context: .
+                  file: ./Dockerfile.enclave
+                  push: false
+                  load: true
+                  tags: ${{ env.ECR_REGISTRY }}/${{ env.ECR_REPOSITORY }}:${{ env.RELEASE_TAG }}
+                  cache-from: type=gha
+                  cache-to: type=gha, mode=max
+
+            - name: Push Docker image to Amazon ECR Public
+              run: |
+                  docker push ${{ env.ECR_REGISTRY }}/${{ env.ECR_REPOSITORY }}:${{ env.RELEASE_TAG }}
+    deploy:
+        needs: docker-build-push
+        runs-on:
+            [
+                "runs-on",
+                "runner=32cpu-linux-x64",
+                "run-id=${{ github.run_id }}",
+                "spot=false",
+            ]
+        steps:
+            - name: Retrieve the tag
+              run: |
+                  # echo "RELEASE_TAG=`git describe --tags --exact-match 2> /dev/null`" >> $GITHUB_ENV
+                  echo "ENCLAVE_VERSION=v1" >> $GITHUB_ENV
+                  echo "RELEASE_TAG=v1" >> $GITHUB_ENV
+
+            - name: Checkout repo
+              uses: actions/checkout@v4
+              with:
+                  fetch-depth: 0
+
+            - name: Setup Node.js
+              uses: actions/setup-node@v4
+              with:
+                  node-version: "20"
+                  cache: "npm"
+
+            - name: Install dependencies
+              run: npm ci
+
+            - name: Install AWS CDK
+              run: npm install -g aws-cdk
+
+            - name: Deploy Base Stack
+              run: |
+                  cdk deploy Sp1TeeBaseStack \
+                    --parameters CertificateArn=${{ secrets.CERTIFICATE_ARN }} \
+                    --parameters PagerDutyWebhookUrl=${{ secrets.PAGERDUTY_WEBHOOK_URL }} \
+                    --require-approval never
+
+            - name: Deploy Versioned Stack
+              run: |
+                  cdk deploy Sp1TeeV${{ env.ENCLAVE_VERSION }}Stack \
+                    --require-approval never
