+++ Dockerfile.enclave
@@ -1,10 +1,7 @@
-ARG ENCLAVE_VERSION
-
 # ---- Build Stage ----
 FROM public.ecr.aws/amazonlinux/amazonlinux:2023 AS builder
 
 ARG DEBUG_MODE=0
-ARG ENCLAVE_VERSION
 
 # Install system dependencies required to build Rust projects
 RUN yum update -y \
@@ -31,10 +28,8 @@ COPY . ./
 RUN cmake --version
 
 RUN if [ "${DEBUG_MODE}" -eq "1" ]; then \
-        cd enclave/v${ENCLAVE_VERSION}; \
         cargo build --release --bin sp1-tee-enclave --features debug-mode,enclave; \
     else \
-        cd enclave/v${ENCLAVE_VERSION}; \
         # Use --locked to ensure the build is reproducible.
         cargo build --release --bin sp1-tee-enclave --locked --features enclave; \
     fi
@@ -42,10 +37,8 @@ RUN if [ "${DEBUG_MODE}" -eq "1" ]; then \
 # ---- Runtime Stage ----
 FROM public.ecr.aws/amazonlinux/amazonlinux:2023
 
-ARG ENCLAVE_VERSION
-
 # Copy the binary from the build stage
-COPY --from=builder /app/enclave/v$ENCLAVE_VERSION/target/release/sp1-tee-enclave /usr/local/bin/sp1-tee-enclave
+COPY --from=builder /app/target/release/sp1-tee-enclave /usr/local/bin/sp1-tee-enclave
 
 # Allow 1 Billion total unconstrained cycles.
 ENV UNCONSTRAINED_CYCLE_LIMIT=1000000000
